Added methods on Set - difference, intersection, is disjoint, is subset, is superset, symmetric_difference, union
The reason these methods are on Observe objects rather than on
Subscribe objects is because PyReactive doesn’t support parentheses as
of yet. By having each operation as a different Observe object, we can
ensure that there’s no ambiguity in the order of precedence, a
situation that is highly likely if the current Subscribe API is used.